GOV-35503 Number-picker validation should allow floats (1203)

+++ d360.web/Scripts/app/components/shared/controls/number-picker/number-input.component.ts
@@ -127,7 +127,7 @@ export class IgNumberFieldcomponent implements ControlValueAccessor, OnInit, Val
 
     validate(control: AbstractControl): ValidationErrors | null {
 		let result: ValidationErrors | null = null;
-        if (isNaN(parseInt(control.value))) {
+        if (isNaN(parseFloat(control.value))) {
             result = null;
         } else if (this.isOverMax()) {
             result = {
